Primer proyecto de la Segunda Evaluación

+++ Ejemplo/Ejemplo/src/Calculadora.java
@@ -1,10 +1,9 @@
 import java.util.Scanner;
 
 public class Calculadora {
-    public static double sumar(int num1, int num2) {
+    public static int sumar(int num1, int num2) {
         return num1 + num2;
     }
-
     public static void restar() {
         Scanner lector = new Scanner(System.in);
         System.out.println("Introduce dos números");
@@ -12,7 +11,6 @@ public class Calculadora {
         int b = lector.nextInt();
         System.out.println(a - b);
     }
-
     public static void producto(int a, int b) {
         System.out.println(a * b);
     }
